All modules related to map was moved to folder 'map'
