+++ js/form/form.js
@@ -11,10 +11,12 @@ import {
   capacity,
   roomNumber
 } from './form-validate-room-capacity.js';
-import { resetMap } from '../map.js';
+import { resetMap } from '../map/map.js';
 
 const advertForm = document.querySelector('.ad-form');
 const formSubmitButton = advertForm.querySelector('.ad-form__submit');
+const timein = advertForm.querySelector('#timein');
+const timeout = advertForm.querySelector('#timeout');
 const formResetButton = advertForm.querySelector('.ad-form__reset');
 
 const callAndAddInputListener = (elem, func) => {
@@ -53,6 +55,14 @@ title.addEventListener('blur', callAndAddInputListener(title, validateTitle));
 price.addEventListener('blur', callAndAddInputListener(price, validatePrice));
 
 type.addEventListener('change', changePricePlaceholderAndMin);
+
+timein.addEventListener('change', () => {
+  timeout.value = timein.value;
+});
+
+timeout.addEventListener('change', () => {
+  timein.value = timeout.value;
+});
 
 formSubmitButton.addEventListener('click', (evt) => {
   const isTitleValid = validateTitle();
