+++ js/map/api.js
@@ -1,5 +1,5 @@
 import { createPopups } from './map.js';
-import { showServerErrorMessage } from './utils.js';
+import { showServerErrorMessage } from '../utils.js';
 
 const DATA_FETCH_LINK = 'https://23.javascript.pages.academy/keksobooking/data';
