+++ js/map/similar-adverts-utils.js
