+++ js/map/similar-adverts.js
